05-agentic-rag-realtime/requirements.txt
langchain==0.3.14
langchain-community==0.3.14
langchain-openai==0.3.0
faiss-cpu==1.9.0.post1
sentence-transformers==3.3.1
pypdf==5.1.0
python-dotenv==1.0.1
huggingface-hub==0.27.1
yfinance==0.2.51
wikipedia==1.4.0
requests==2.32.3
